refactor: cleaned file imports in index.css
