+++ public/chunk-map.json
@@ -1,1 +1,1 @@
-{"polyfill":["/polyfill.js","/polyfill.09ef23fc523d3762efcf.hot-update.js"],"commons":["/commons.css","/commons.js","/commons.09ef23fc523d3762efcf.hot-update.js"],"component---cache-dev-404-page-js":["/component---cache-dev-404-page-js.js"],"component---src-pages-404-js":["/component---src-pages-404-js.js"],"component---src-pages-about-js":["/component---src-pages-about-js.js"],"component---src-pages-contact-js":["/component---src-pages-contact-js.js"],"component---src-pages-index-js":["/component---src-pages-index-js.js","/component---src-pages-index-js.09ef23fc523d3762efcf.hot-update.js"],"component---src-pages-work-js":["/component---src-pages-work-js.js"],"component---src-templates-work-project-work-js":["/component---src-templates-work-project-work-js.js"]}
+{"polyfill":["/polyfill.js","/polyfill.5d4829776d80a5436950.hot-update.js"],"commons":["/commons.css","/commons.js","/commons.5d4829776d80a5436950.hot-update.js"],"component---cache-dev-404-page-js":["/component---cache-dev-404-page-js.js"],"component---src-pages-404-js":["/component---src-pages-404-js.js","/component---src-pages-404-js.5d4829776d80a5436950.hot-update.js"],"component---src-pages-about-js":["/component---src-pages-about-js.js"],"component---src-pages-contact-js":["/component---src-pages-contact-js.js"],"component---src-pages-index-js":["/component---src-pages-index-js.js"],"component---src-pages-work-js":["/component---src-pages-work-js.js"],"component---src-templates-work-project-work-js":["/component---src-templates-work-project-work-js.js"]}
